// UIL/src/app/shared/models/gameonruninfo.model.ts
export interface GameOnRunModel
{
    GameId: number;

    FirstPlayerFrontRight: number;

    FirstPlayerFrontLeft: number;

    FirstPlayerBackRight: number;

    FirstPlayerBackLeft: number;

    SecondPlayerFrontRight: number;

    SecondPlayerFrontLeft: number;

    SecondPlayerBackRight: number;

    SecondPlayerBackLeft: number;

    UsedCard: number;
}